Add build environment for slim lambda dependencies
Creates an container on which amazonlinux based libraries can be
compiled for python dependencies targeting a lambda execution
environment.  Small refactor of the png tile generation to make it
possible to selectively load files without requiring PIL
