+++ docker-compose.yml
@@ -15,3 +15,14 @@ services:
       - $HOME/.aws:/root/.aws
     environment:
       AWS_PROFILE: "azavea-datahub"
+  lambda:
+    image: srp-lambda-geop
+    build:
+      context: ./src
+      dockerfile: Dockerfile.lambda
+    volumes:
+      - ./src/geop:/usr/src/geop
+      - ./src/test_data:/usr/src/test_data
+      - $HOME/.aws:/root/.aws
+    environment:
+      AWS_PROFILE: "azavea-datahub"
